Correction des migrations

+++ database/migrations/2022_10_25_120426_add_description_id_to_offres_table.php
@@ -25,8 +25,11 @@ class AddDescriptionIdToOffresTable extends Migration
      */
     public function down()
     {
-        Schema::table('offres', function (Blueprint $table) {
-            $table->dropColumn('description');
-        });
+        if (Schema::hasColumn('offres', 'description')) {
+            Schema::table('offres', function (Blueprint $table)
+            {
+                $table->dropColumn('description');
+            });
+        }
     }
 }
